Move playback service to module
