+++ app/src/main/java/de/danoeh/antennapod/dialog/StreamingConfirmationDialog.java
@@ -3,9 +3,9 @@ package de.danoeh.antennapod.dialog;
 import android.content.Context;
 import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 import de.danoeh.antennapod.R;
+import de.danoeh.antennapod.playback.service.PlaybackServiceStarter;
 import de.danoeh.antennapod.storage.preferences.UserPreferences;
 import de.danoeh.antennapod.model.playback.Playable;
-import de.danoeh.antennapod.core.util.playback.PlaybackServiceStarter;
 
 public class StreamingConfirmationDialog {
     private final Context context;
