+++ app/src/main/java/de/danoeh/antennapod/dialog/VariableSpeedDialog.java
@@ -17,8 +17,8 @@ import com.google.android.material.chip.Chip;
 import com.google.android.material.snackbar.Snackbar;
 import de.danoeh.antennapod.R;
 import de.danoeh.antennapod.event.playback.SpeedChangedEvent;
+import de.danoeh.antennapod.playback.service.PlaybackController;
 import de.danoeh.antennapod.storage.preferences.UserPreferences;
-import de.danoeh.antennapod.core.util.playback.PlaybackController;
 import de.danoeh.antennapod.view.ItemOffsetDecoration;
 import de.danoeh.antennapod.view.PlaybackSpeedSeekBar;
 import org.greenrobot.eventbus.EventBus;
