+++ config/spotbugs/exclude.xml
@@ -30,7 +30,7 @@
     </Match>
     <Match>
         <Bug pattern="MS_CANNOT_BE_FINAL"/>
-        <Class name="de.danoeh.antennapod.core.service.playback.PlaybackService"/>
+        <Class name="de.danoeh.antennapod.playback.service.PlaybackService"/>
     </Match>
     <Match>
         <Bug pattern="MS_MUTABLE_ARRAY"/>
@@ -62,7 +62,7 @@
     </Match>
     <Match>
         <Bug pattern="ST_WRITE_TO_STATIC_FROM_INSTANCE_METHOD"/>
-        <Class name="de.danoeh.antennapod.core.service.playback.PlaybackService"/>
+        <Class name="de.danoeh.antennapod.playback.service.PlaybackService"/>
     </Match>
     <Match>
         <Bug pattern="ST_WRITE_TO_STATIC_FROM_INSTANCE_METHOD"/>
