+++ core/src/main/AndroidManifest.xml
@@ -12,60 +12,4 @@
     <uses-permission android:name="android.permission.VIBRATE" />
     <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
 
-    <application
-        android:allowBackup="true"
-        android:icon="@mipmap/ic_launcher"
-        android:supportsRtl="true">
-
-        <service android:name=".service.playback.PlaybackService"
-            android:label="@string/app_name"
-            android:enabled="true"
-            android:exported="true"
-            android:foregroundServiceType="mediaPlayback"
-            tools:ignore="ExportedService">
-
-            <intent-filter>
-                <action android:name="android.media.browse.MediaBrowserService"/>
-                <action android:name="de.danoeh.antennapod.intents.PLAYBACK_SERVICE" />
-            </intent-filter>
-        </service>
-
-        <receiver
-            android:name=".receiver.MediaButtonReceiver"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MEDIA_BUTTON" />
-            </intent-filter>
-            <intent-filter>
-                <action android:name="de.danoeh.antennapod.NOTIFY_BUTTON_RECEIVER" />
-            </intent-filter>
-        </receiver>
-
-        <receiver android:name=".receiver.FeedUpdateReceiver"
-            android:label="@string/feed_update_receiver_name"
-            android:exported="true"
-            tools:ignore="ExportedReceiver" /> <!-- allow feeds update to be triggered by external apps -->
-
-        <service
-            android:name=".service.QuickSettingsTileService"
-            android:enabled="true"
-            android:exported="true"
-            android:label="@string/app_name"
-            android:icon="@drawable/ic_notification"
-            android:permission="android.permission.BIND_QUICK_SETTINGS_TILE">
-            <intent-filter>
-                <action android:name="android.service.quicksettings.action.QS_TILE" />
-            </intent-filter>
-            <meta-data android:name="android.service.quicksettings.ACTIVE_TILE" android:value="true" />
-            <meta-data android:name="android.service.quicksettings.TOGGLEABLE_TILE" android:value="true" />
-        </service>
-    </application>
-
-    <queries>
-        <intent>
-            <action android:name="android.intent.action.VIEW" />
-            <data android:scheme="https" />
-        </intent>
-    </queries>
-
 </manifest>
