+++ core/src/main/res/values/ids.xml
@@ -14,8 +14,4 @@
 
     <!-- View types -->
     <item name="view_type_episode_item" type="id"/>
-
-    <!-- Notifications need unique IDs to update/cancel them -->
-    <item name="notification_playing" type="id"/>
-    <item name="notification_streaming_confirmation" type="id"/>
 </resources>
