+++ net/download/service/src/main/AndroidManifest.xml
@@ -1,7 +1,19 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" xmlns:tools="http://schemas.android.com/tools">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-permission android:name="android.permission.ACCESS_WIFI_STATE" />
     <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
+
+    <application
+        android:allowBackup="true"
+        android:icon="@mipmap/ic_launcher"
+        android:supportsRtl="true">
+
+        <receiver android:name=".feed.FeedUpdateReceiver"
+                  android:label="@string/feed_update_receiver_name"
+                  android:exported="true"
+                  tools:ignore="ExportedReceiver" /> <!-- allow feeds update to be triggered by external apps -->
+    </application>
+
 </manifest>
