+++ playback/service/README.md
@@ -1,0 +1,3 @@
+# :playback:service
+
+The main service doing media playback.
