+++ playback/service/src/main/java/de/danoeh/antennapod/playback/service/MediaButtonReceiver.java
@@ -1,4 +1,4 @@
-package de.danoeh.antennapod.core.receiver;
+package de.danoeh.antennapod.playback.service;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
